Split out module processor just for GlideModules.
