+++ annotation/compiler/src/main/java/com/bumptech/glide/annotation/compiler/GlideAnnotationProcessor.java
@@ -1,0 +1,102 @@
+package com.bumptech.glide.annotation.compiler;
+
+import com.google.auto.service.AutoService;
+import java.util.HashSet;
+import java.util.Set;
+import javax.annotation.processing.AbstractProcessor;
+import javax.annotation.processing.ProcessingEnvironment;
+import javax.annotation.processing.Processor;
+import javax.annotation.processing.RoundEnvironment;
+import javax.lang.model.SourceVersion;
+import javax.lang.model.element.TypeElement;
+
+// Links in Javadoc will work due to build setup, even though there is no direct dependency here.
+/**
+ * Generates classes to allow Glide to discover and call all
+ * {@link com.bumptech.glide.module.ChildGlideModule}s in an application and any dependent libraries
+ * of the application and the RootGlideModule included in the application itself.
+ *
+ * <p>This processor discovers all {@link com.bumptech.glide.module.RootGlideModule} and
+ * {@link com.bumptech.glide.module.ChildGlideModule} implementations that are
+ * annotated with {@link com.bumptech.glide.annotation.GlideModule}. Any implementations missing the
+ * annotation will be ignored.
+ *
+ * <p>Two classes are generated by this processor:
+ * <ul>
+ *   <li>For {@link com.bumptech.glide.module.ChildGlideModule}s - A GlideIndexer class in a
+ *      specific package that will later be used by the processor to discover all
+ *      {@link com.bumptech.glide.module.ChildGlideModule} classes.
+ *   <li>For {@link com.bumptech.glide.module.RootGlideModule}s - A single
+ *      {@link com.bumptech.glide.module.RootGlideModule} implementation
+ *     ({@link com.bumptech.glide.GeneratedRootGlideModule}) that calls all
+ *     {@link com.bumptech.glide.module.ChildGlideModule}s and the
+ *     original {@link com.bumptech.glide.module.RootGlideModule} in the correct order when Glide is
+ *     initialized.
+ * </ul>
+ *
+ * <p>{@link com.bumptech.glide.module.RootGlideModule} implementations must only be included in
+ * applications, not in libraries. There must be exactly one
+ * {@link com.bumptech.glide.module.RootGlideModule} implementation per
+ * Application. The {@link com.bumptech.glide.module.RootGlideModule} class is
+ * used as a signal that all modules have been found and that the final merged
+ * {@link com.bumptech.glide.GeneratedRootGlideModule} impl can be created.
+ */
+@AutoService(Processor.class)
+public final class GlideAnnotationProcessor extends AbstractProcessor {
+  static final boolean DEBUG = false;
+  private ProcessorUtil processorUtil;
+  private ChildModuleProcessor childModuleProcessor;
+  private RootModuleProcessor rootModuleProcessor;
+  private boolean isGeneratedRootGlideModuleWritten;
+
+  @Override
+  public synchronized void init(ProcessingEnvironment processingEnvironment) {
+    super.init(processingEnvironment);
+    processorUtil = new ProcessorUtil(processingEnvironment);
+    childModuleProcessor = new ChildModuleProcessor(processorUtil);
+    rootModuleProcessor = new RootModuleProcessor(processingEnvironment, processorUtil);
+  }
+
+  @Override
+  public Set<String> getSupportedAnnotationTypes() {
+    Set<String> result = new HashSet<>();
+    result.addAll(childModuleProcessor.getSupportedAnnotationTypes());
+    return result;
+  }
+
+  @Override
+  public SourceVersion getSupportedSourceVersion() {
+    return SourceVersion.RELEASE_7;
+  }
+
+   /**
+   * Each round we do the following:
+   * <ol>
+   *   <li>Find all RootGlideModules and save them to an instance variable (throw if > 1).
+   *   <li>Find all ChildGlideModules
+   *   <li>For each ChildGlideModule, write an Indexer with an Annotation with the class name.
+   *   <li>If we wrote any Indexers, return and wait for the next round.
+   *   <li>If we didn't write any Indexers and there is a RootGlideModule, write the
+   *   GeneratedRootGlideModule. Once the GeneratedRootGlideModule is written, we expect to be
+   *   finished. Any further generation of related classes will result in errors.
+   * </ol>
+   */
+  @Override
+  public boolean process(Set<? extends TypeElement> set, RoundEnvironment env) {
+    processorUtil.process();
+    boolean newModulesWritten = childModuleProcessor.processModules(set, env);
+    rootModuleProcessor.processModules(set, env);
+
+    if (newModulesWritten) {
+      if (isGeneratedRootGlideModuleWritten) {
+        throw new IllegalStateException("Cannot process annotations after writing RootGlideModule");
+      }
+      return true;
+    }
+
+    if (!isGeneratedRootGlideModuleWritten) {
+      isGeneratedRootGlideModuleWritten = rootModuleProcessor.maybeWriteRootModule();
+    }
+    return true;
+  }
+}
